Hotfix: Make MME address configurable in HSS
In dynamic environments the operator might want to be able to set the
DNS name of the MME and not hardcode it to be mme.localdomain. Until
this "//TODO get the list of peers from the database" is implemented we
should be able to atleast set it through the environment.

MME_IDENTITY env variable needs to be set before running hss to use anything
other than mme.localdomain. Current behavior should continue if this
variable is not set.

Signed-off-by: Saikrishna Edupuganti <saikrishna.edupuganti@intel.com>